requirements.txt
python-dotenv==1.0.0
slack-sdk==3.21.3
html2text==2020.1.16